// src/app/modules/transactions/models/picking-sales-order.ts
export interface PickingSalesOrderRoot {
  header: PickingSalesOrderHeader
  details: PickingSalesOrderDetail[]
  pickingHistory: any[]
}

export interface PickingSalesOrderHeader {
  salesOrderId: number
  salesOrderNum: string
  trxDate: string
  customerId: number
  businessModelType: string
  countryId: number
  currencyId: number
  typeCode: string
  locationId: number
  toLocationId: any
  location: string
  toLocation: any
  deactivated: any
}

export interface PickingSalesOrderDetail {
  salesOrderId: number
  itemId: number
  description: string
  itemVariationXId: number
  itemVariationYId?: number
  itemSku: string
  itemVariationTypeCode: any
  itemCode: string
  itemVariationXDescription: string
  itemVariationYDescription?: string
  itemUomId: number
  itemUomDescription: string
  rack?: string
  subRack?: string
  qtyRequest: number
  qtyCommit: number
  qtyBalance: any
  qtyPicked: number
  qtyPickedCurrent: number
  qtyPacked: number
}